fix: Change default of onlyWithURL, remove graceful-fs, modernise sources
* Trying to load source maps for files without `sourceMappingURL`
  is not something usual and should not be the default behaviour.
* Windows support in Node.js has improved and graceful-fs should
  not be needed any more.
* Use modern JavaScript language syntax. Node.js 15 is required
  anyway, at least because of the need for `replaceAll`.

BREAKING CHANGE: The package `graceful-fs` is not used any more. It should not affect your project, but better to test it. Source maps are loaded only for files with `sourceMappingURL` by default. (The default value of the option `onlyWithURL` has changed to `true`.) If your JavaScript files lack `sourceMappingURL`, but include external source maps, set the options `onlyWithURL` to `false`.